requirements.txt
git+git://github.com/tarnacious/mitmproxy.git@master#egg=mitmproxy
aioredis==1.3.1
rratelimit==0.0.4
redis==2.10.5
cssmin==0.2.0
gunicorn==19.3.0
psycopg2==2.7.5
Flask==0.12
Flask-Login==0.4.0
Flask-Migrate==2.0.3
Flask-SQLAlchemy==2.2
Flask-WTF==0.14
Flask-User==0.6.9
pytest==3.0.2
pytest-flask==0.8.1
pytest-cov==1.8.1
pytest-capturelog
testfixtures
webassets==0.12.1
Flask-Assets==0.12
mypy==0.471
requests==2.17.3
freezegun==0.3.9
concurrent-log-handler==0.9.7
ansible==2.6.3
